moviebackend/requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-cors-headers==3.5.0
django-import-export==2.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
numpy==1.19.2
odfpy==1.4.1
opencv-python==4.4.0.44
openpyxl==3.0.4
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
